update file_manager.rs FileManager struct

+++ src/managers/file_manager.rs
@@ -10,14 +10,16 @@ use crate::managers::manager::Manager;
 pub struct FileManager {
     config: PathBuf,
     locker: PathBuf,
+    index: PathBuf
 }
 
 impl FileManager {
-    pub fn new(config: &PathBuf, locker: &PathBuf) -> FileManager {
+    pub fn new(index: &PathBuf, config: &PathBuf, locker: &PathBuf) -> FileManager {
+        let index = index.clone();
         let config = config.clone();
         let locker = locker.clone();
 
-        let mut fm = FileManager { config, locker };
+        let mut fm = FileManager { config, locker, index };
 
         fm.init().expect("Could not initialize FileManager");
 
@@ -60,22 +62,17 @@ impl FileManager {
         )
     }
 
-    // NOTE: this can be moved to the Manager Trait
-    // if field traits are implemented in the future
     fn gen_path(&self, for_path: &str, path: &str) -> String {
         let mut location = PathBuf::new();
 
         match for_path {
-            "locker" => {
-                location.push(self.locker.clone());
-                location.push(path);
-            },
-            "config" => {
-                location.push(self.config.clone());
-                location.push(path);
-            },
+            "index" => { location.push(self.index.clone()); },
+            "locker" => { location.push(self.locker.clone()); },
+            "config" => { location.push(self.config.clone()); },
             _ => panic!("Unsupported location {:?}", location.as_path().to_str())
         };
+
+        location.push(path);
 
         Self::pb_to_str(&location)
     }
@@ -211,7 +208,16 @@ impl Manager for FileManager {
     // the default settings.yml file is
     // required on the config_path
     fn init(&mut self) -> io::Result<()> {
+        let index_path = self.index.as_path().to_owned();
         let config_path = self.config.as_path().to_owned();
+        
+        if !index_path.exists() {
+            self.create(
+                index_path
+                    .to_str()
+                    .unwrap()
+            )?;
+        }
 
         if !config_path.exists() { 
             self.create(
@@ -227,7 +233,10 @@ impl Manager for FileManager {
     fn create(&mut self, path: &str) -> io::Result<()> {
         let p = Path::new(path);
 
-        if !p.exists() { File::create(p).expect("Unable to create file"); }
+        if !p.exists() { 
+            File::create(p)
+                .expect(&format!("Unable to create file {:?}", p)); 
+        }
 
         Ok(()) 
     }
@@ -235,7 +244,11 @@ impl Manager for FileManager {
     fn read(&mut self, path: &str) -> io::Result<Self::Output> {
         let p = Path::new(path);
 
-        if !p.exists() { panic!("Trying to open file that does not exist"); }
+        if !p.exists() {
+            panic!(
+                format!("Trying to open file {:?} that does not exist", path)
+            ); 
+        }
 
         let mut file = File::open(p)?;
         let mut contents = String::new();
@@ -267,12 +280,9 @@ mod test {
     fn after_each(this: &mut Setup) {
         for path in this.paths.iter() {
             let exists = Path::new(&path).exists();
+            let msg = format!("Could not remove {} in `file_manager.rs` test", path);
 
-            if exists {
-                let msg = format!("Could not remove {} in `file_manager.rs` test", path);
-                
-                remove_file(path).expect(&msg);
-            }
+            if exists { remove_file(path).expect(&msg); }
         }
     }
 
@@ -282,8 +292,8 @@ mod test {
             paths: Vec::new(),
             after_each: &after_each,
             test: &|this| {
-                let (config, locker) = this.as_path_buf();
-                FileManager::new(&config, &locker);
+                let (index, config, locker) = this.as_path_buf();
+                FileManager::new(&index, &config, &locker);
             }
         }; 
     }
@@ -294,9 +304,9 @@ mod test {
             paths: Vec::new(),
             after_each: &after_each,
             test: &|this| {
-                let (config, mut locker) = this.as_path_buf();
+                let (index, config, mut locker) = this.as_path_buf();
                 
-                let mut fm = FileManager::new(&config, &locker);
+                let mut fm = FileManager::new(&index, &config, &locker);
                 let locker_path = FileManager::pb_to_str(&locker); 
 
                 locker.push(&locker_path);
@@ -313,9 +323,9 @@ mod test {
             paths: Vec::new(),
             after_each: &after_each,
             test: &|this| {
-                let (mut config, locker) = this.as_path_buf();
+                let (index, mut config, locker) = this.as_path_buf();
                 
-                let mut fm = FileManager::new(&config, &locker);
+                let mut fm = FileManager::new(&index, &config, &locker);
                 let config_path = FileManager::pb_to_str(&config);
 
                 config.push(&config_path);
@@ -332,9 +342,9 @@ mod test {
             paths: Vec::new(),
             after_each: &after_each,
             test: &|this| {
-                let (config, mut locker) = this.as_path_buf();
+                let (index, config, mut locker) = this.as_path_buf();
                 
-                let mut fm = FileManager::new(&config, &locker);
+                let mut fm = FileManager::new(&index, &config, &locker);
                 let locker_path = FileManager::pb_to_str(&locker); 
 
                 fm.create_locker(&locker_path);
@@ -352,11 +362,12 @@ mod test {
             paths: Vec::new(),
             after_each: &after_each,
             test: &|this| {
-                let (config, locker) = this.as_path_buf();
-                let config_path = FileManager::pb_to_str(&config);
-                let mut fm = FileManager::new(&config, &locker);
+                let (index, config, locker) = this.as_path_buf();
                 
-                let file = fm.read_locker(&config_path).unwrap();
+                let config_path = FileManager::pb_to_str(&config);
+                let mut fm = FileManager::new(&index, &config, &locker);
+                
+                let file = fm.read_config(&config_path).unwrap();
 
                 assert_eq!(file, String::from(""));
             }
@@ -369,8 +380,8 @@ mod test {
             paths: Vec::new(),
             after_each: &after_each,
             test: &|this| {
-                let (config, locker) = this.as_path_buf();
-                let mut fm = FileManager::new(&config, &locker);
+                let (index, config, locker) = this.as_path_buf();
+                let mut fm = FileManager::new(&index, &config, &locker);
 
                 // https://doc.rust-lang.org/std/panic/struct.AssertUnwindSafe.html
                 let result = catch_unwind(AssertUnwindSafe(|| {
@@ -389,8 +400,8 @@ mod test {
             paths: Vec::new(),
             after_each: &after_each,
             test: &|this| {
-                let (config, locker) = this.as_path_buf();
-                let mut fm = FileManager::new(&config, &locker);
+                let (index, config, locker) = this.as_path_buf();
+                let mut fm = FileManager::new(&index, &config, &locker);
 
                 // https://doc.rust-lang.org/std/panic/struct.AssertUnwindSafe.html
                 let result = catch_unwind(AssertUnwindSafe(|| {
@@ -409,8 +420,8 @@ mod test {
             paths: Vec::new(),
             after_each: &after_each,
             test: &|this| {
-                let (config, locker) = this.as_path_buf();
-                let mut fm = FileManager::new(&config, &locker);
+                let (index, config, locker) = this.as_path_buf();
+                let mut fm = FileManager::new(&index, &config, &locker);
                 let path_to_remove = FileManager::pb_to_str(&locker);
                 
                 fm.remove_locker(&path_to_remove);
@@ -427,8 +438,8 @@ mod test {
             paths: Vec::new(),
             after_each: &after_each,
             test: &|this| {
-                let (config, locker) = this.as_path_buf();
-                let mut fm = FileManager::new(&config, &locker);
+                let (index, config, locker) = this.as_path_buf();
+                let mut fm = FileManager::new(&index, &config, &locker);
                 let path_to_remove = FileManager::pb_to_str(&config);
                 
                 fm.remove_config(&path_to_remove);
